+++ src/main/java/com/pricewagon/pricewagon/domain/product/repository/ProductHistoryCustomRepository.java
@@ -5,5 +5,5 @@ import java.util.Optional;
 import com.pricewagon.pricewagon.domain.product.entity.ProductHistory;
 
 public interface ProductHistoryCustomRepository {
-	Optional<ProductHistory> findFirstByProductIdAndPriceNot(Long productId, Integer currentPrice);
+	Optional<ProductHistory> findFirstByProductIdAndPriceNot(Integer productId, Integer currentPrice);
 }
