+++ src/main/java/com/pricewagon/pricewagon/domain/product/repository/ProductHistoryRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pricewagon.pricewagon.domain.product.entity.ProductHistory;
 
-public interface ProductHistoryRepository extends JpaRepository<ProductHistory, Long> , ProductHistoryCustomRepository{
+public interface ProductHistoryRepository extends JpaRepository<ProductHistory, Integer> , ProductHistoryCustomRepository{
 
 
 }
